+++ lib/src/shim/dart_ui_web_fake.dart
@@ -11,6 +11,5 @@ class PlatformViewRegistry {
     String viewType,
     Function viewFactory, {
     bool isVisible = true,
-  }) =>
-      false;
+  }) => false;
 }