+++ test/html_builder_test.dart
@@ -6,7 +6,7 @@ import 'package:model_viewer_plus/src/html_builder.dart' show HTMLBuilder;
 void main() {
   group('HTMLBuilder', () {
     test('supports the src and backgroundColor attributes', () {
-      final html = HTMLBuilder.build(
+      final String html = HTMLBuilder.build(
         src: 'src.glb',
         htmlTemplate: '<!-- body -->',
         backgroundColor: const Color(0xFFABCDEF),